move StoreRegistry to pro directory
