+++ node_package/src/ReactOnRails.client.ts
@@ -2,7 +2,7 @@ import type { ReactElement } from 'react';
 import * as ClientStartup from './clientStartup.ts';
 import { renderOrHydrateComponent, hydrateStore } from './ClientSideRenderer.ts';
 import * as ComponentRegistry from './pro/ComponentRegistry.ts';
-import * as StoreRegistry from './StoreRegistry.ts';
+import * as StoreRegistry from './pro/StoreRegistry.ts';
 import buildConsoleReplay from './buildConsoleReplay.ts';
 import createReactOutput from './createReactOutput.ts';
 import * as Authenticity from './Authenticity.ts';
